chore(team): split mailbox shared thread helpers
