+++ src/team/manager.rs
@@ -8,6 +8,7 @@ mod mailbox;
 mod mailbox_payloads;
 mod mailbox_reply_obligations;
 mod mailbox_service;
+mod mailbox_shared_thread;
 mod mailbox_store;
 mod mailbox_threads;
 mod memory_flush;
